+++ tensorflow_quantum/core/ops/math_ops/tfq_inner_product_grad.cc
@@ -332,7 +332,7 @@ class TfqInnerProductGradOp : public tensorflow::OpKernel {
 
     const int output_dim_internal_size = other_fused_circuits[0].size();
 
-    auto DoWork = [&](int start, int end) {
+    auto DoWork = [&](int64_t start, int64_t end) {
       int old_batch_index = -2;
       int cur_batch_index = -1;
       uint64_t largest_nq = 1;
@@ -344,7 +344,7 @@ class TfqInnerProductGradOp : public tensorflow::OpKernel {
       auto sv_adj = ss.Create(largest_nq);
       auto scratch = ss.Create(largest_nq);
       auto scratch2 = ss.Create(largest_nq);
-      for (int i = start; i < end; i++) {
+      for (int64_t i = start; i < end; i++) {
         cur_batch_index = i / output_dim_internal_size;
         cur_internal_index = i % output_dim_internal_size;
